+++ examples/main/modules.tf
@@ -15,6 +15,18 @@ module "rg" {
   stack       = var.stack
 }
 
+module "logs" {
+  source  = "claranet/run-common/azurerm//modules/logs"
+  version = "x.x.x"
+
+  client_name         = var.client_name
+  environment         = var.environment
+  stack               = var.stack
+  location            = module.azure_region.location
+  location_short      = module.azure_region.location_short
+  resource_group_name = module.rg.resource_group_name
+}
+
 module "apim" {
   source  = "claranet/api-management/azurerm"
   version = "x.x.x"
@@ -49,5 +61,10 @@ module "apim" {
       location  = "eastus2"
       subnet_id = var.subnet_id
     },
+  ]
+
+  logs_destinations_ids = [
+    module.logs.logs_storage_account_id,
+    module.logs.log_analytics_workspace_id
   ]
 }