+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Breaking
   * AZ-589: Upgrade `diagnostics` to `v5.0.0`
   * AZ-589: Variables name cleanup
+  * AZ-515: Require Terraform 0.13+
 
 Added
   * AZ-615: Add an option to enable or disable default tags